pyproject.toml
[project]
name = "music-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "celery>=5.5.3",
    "email-validator>=2.3.0",
    "fastapi-cache>=0.1.0",
    "fastapi[standart]>=0.117.1",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "orjson>=3.11.3",
    "pip-system-certs>=5.2",
    "pydantic-settings>=2.11.0",
    "pyjwt==2.8.*",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
]
